# Add Semi Sorted Batching.