--- a/nemo/core/optim/lr_scheduler.py
+++ b/nemo/core/optim/lr_scheduler.py
@@ -878,7 +878,10 @@ def prepare_lr_scheduler(
             else:
                 raise ValueError(f'Could not find batch_size from batch_sampler: {train_dataloader.batch_sampler}')
         elif hasattr(train_dataloader, 'sampler') and train_dataloader.sampler is not None:
-            if hasattr(train_dataloader.sampler, 'micro_batch_size') and train_dataloader.sampler.micro_batch_size is not None:
+            if (
+                hasattr(train_dataloader.sampler, 'micro_batch_size')
+                and train_dataloader.sampler.micro_batch_size is not None
+            ):
                 batch_size = train_dataloader.sampler.micro_batch_size
             else:
                 raise ValueError(f'Could not find batch_size from sampler: {train_dataloader.sampler}')